Deprecate `UIView.AngleUnit`, `UIView.rotate(byAngle:ofType:animated:duration:completion:)` and `UIView.rotate(toAngle:ofType:animated:duration:completion:)`
