+++ Sources/SwifterSwift/UIKit/UIViewExtensions.swift
@@ -18,18 +18,6 @@ public extension UIView {
         case vertical
     }
 
-    /// SwifterSwift: Angle units.
-    ///
-    /// - degrees: degrees.
-    /// - radians: radians.
-    enum AngleUnit {
-        /// SwifterSwift: degrees.
-        case degrees
-
-        /// SwifterSwift: radians.
-        case radians
-    }
-
     /// SwifterSwift: Shake animations types.
     ///
     /// - linear: linear animation.
@@ -387,48 +375,6 @@ public extension UIView {
         let radians = CGFloat(angle.converted(to: .radians).value)
         UIView.animate(withDuration: aDuration, delay: 0, options: .curveLinear, animations: { () -> Void in
             self.transform = self.transform.rotated(by: radians)
-        }, completion: completion)
-    }
-
-    /// SwifterSwift: Rotate view by angle on relative axis.
-    ///
-    /// - Parameters:
-    ///   - angle: angle to rotate view by.
-    ///   - type: type of the rotation angle.
-    ///   - animated: set true to animate rotation (default is true).
-    ///   - duration: animation duration in seconds (default is 1 second).
-    ///   - completion: optional completion handler to run with animation finishes (default is nil).
-    func rotate(
-        byAngle angle: CGFloat,
-        ofType type: AngleUnit,
-        animated: Bool = false,
-        duration: TimeInterval = 1,
-        completion: ((Bool) -> Void)? = nil) {
-        let angleWithType = (type == .degrees) ? .pi * angle / 180.0 : angle
-        let aDuration = animated ? duration : 0
-        UIView.animate(withDuration: aDuration, delay: 0, options: .curveLinear, animations: { () -> Void in
-            self.transform = self.transform.rotated(by: angleWithType)
-        }, completion: completion)
-    }
-
-    /// SwifterSwift: Rotate view to angle on fixed axis.
-    ///
-    /// - Parameters:
-    ///   - angle: angle to rotate view to.
-    ///   - type: type of the rotation angle.
-    ///   - animated: set true to animate rotation (default is false).
-    ///   - duration: animation duration in seconds (default is 1 second).
-    ///   - completion: optional completion handler to run with animation finishes (default is nil).
-    func rotate(
-        toAngle angle: CGFloat,
-        ofType type: AngleUnit,
-        animated: Bool = false,
-        duration: TimeInterval = 1,
-        completion: ((Bool) -> Void)? = nil) {
-        let angleWithType = (type == .degrees) ? .pi * angle / 180.0 : angle
-        let aDuration = animated ? duration : 0
-        UIView.animate(withDuration: aDuration, animations: {
-            self.transform = self.transform.concatenating(CGAffineTransform(rotationAngle: angleWithType))
         }, completion: completion)
     }
